+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Below are changes from 7.0.46 to 7.1.0 :
 ## New features
 
 -   You can now open a song's media file from the song list in system panel (#1454)
+-   Added fallback playlist KM can play when your current playlist has ended (#1418)
 -   Display year and duration of song in version selector in public interface (#1448)
 -   Prevent app from starting up if it detects an older database version (#1427)
 -   Added Mastodon link (#1421)